live fixes

+++ grasp_planning_graspit/om/autograsp_planning.cpp
@@ -39,7 +39,7 @@ double * quickGrasp(
     std::string& objectFilename, 
     std::string& robotFilename,
     Eigen::Vector3d& robPos, 
-    std::vector<float>& robRot)
+    std::vector<double>& robRot)
 {
     // PRINTMSG("Initializing GraspIt")
     SHARED_PTR<GraspIt::GraspItSceneManager> graspitMgr(new GraspIt::GraspItSceneManagerHeadless());
